requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
noise
numpy